feat(web): added notify Error utility function
