+++ apps/web/app/HeaderLayout.tsx
@@ -9,13 +9,12 @@ import {
 	TextInput
 } from '@mantine/core'
 import { useQuery } from '@tanstack/react-query'
-import { AxiosError } from 'axios'
 import NextLink from 'next/link'
 import { useRouter } from 'next/navigation'
-import { ReactNode } from 'react'
+import { ReactNode, useEffect as useFootgun } from 'react'
 import { $Api } from '../libs'
-import { Notification, useNotification } from '../stores'
-import { uuid } from '../utils'
+import { useNotification } from '../stores'
+import { notifyError } from '../utils'
 import ArrowUpRightFromSquareIcon from './Icons/ArrowUpRightFromSquareIcon.svg'
 import BellIcon from './Icons/BellIcon.svg'
 import CaretDownIcon from './Icons/CaretDownIcon.svg'
@@ -45,26 +44,16 @@ export default function HeaderLayout({ children }: { children: ReactNode }) {
 		queryFn: $Api.auth.authMeGet
 	})
 
-	if (error && error instanceof AxiosError) {
-		const notification: Notification = {
-			type: 'error',
-			k: uuid()
-		}
-		if (error.response) {
-			notification.title = error.response.data
-		} else if (error.message) {
-			notification.title = error.message
-		}
-
-		notify(notification)
-	}
-
 	const handleLogout = () => {
 		localStorage.removeItem('accessToken')
 		localStorage.removeItem('refreshToken')
 
 		router.push('/login')
 	}
+
+	useFootgun(() => {
+		notifyError(error, notify)
+	}, [error])
 
 	return (
 		<>
